Adding token validation for gets and posts, fixing CSRF Vulnerabilities

+++ includes/model.php
@@ -17,36 +17,43 @@ class model {
 	 * Prepares a INSERT query to the database
 	 */
 	public function save($attributes) {
-		global $db; 
-		$new_attributes = $this->filterParams($attributes, $this->getColumns());
-		$sql = 'INSERT INTO '.$this->getTableName().' '.
-				'('.join(', ',array_keys($new_attributes)).') '.
-				'VALUES ('.join(',',array_values($new_attributes)).')';
-		$db->query($sql);
-	    return $db->insert_id();
+		global $main, $db; 
+		if ($main->checkToken()) {	
+			$new_attributes = $this->filterParams($attributes, $this->getColumns());
+			$sql = 'INSERT INTO '.$this->getTableName().' '.
+					'('.join(', ',array_keys($new_attributes)).') '.
+					'VALUES ('.join(',',array_values($new_attributes)).')';
+			$db->query($sql);
+			return $db->insert_id();
+		}
+		return false;	    
 	}
 	
 	/**
 	 * Builds an update query to hit the DB
 	 */
 	public function update($attributes) {
-		global $db;
-		//Remove the primary key id
-		unset($attributes['id']);		
-		$sql = 'UPDATE '.$this->getTableName().' '.
-        		'SET '.join(', ', $this->getAvailableAttributesQuoted($attributes)) .' '.
-        		"WHERE id ='".$this->getPrimaryKey()."'";
-       	$db->query($sql); 
+		global $main, $db;
+		if ($main->checkToken()) {	
+			//Remove the primary key id
+			unset($attributes['id']);		
+			$sql = 'UPDATE '.$this->getTableName().' '.
+	        		'SET '.join(', ', $this->getAvailableAttributesQuoted($attributes)) .' '.
+	        		"WHERE id ='".$this->getPrimaryKey()."'";
+	       	$db->query($sql); 
+		}
 	}
 	
 	/**
 	 * Deletes a record
 	 */
 	public function delete() {
-		global $db;
-		$sql = 'DELETE FROM '.$this->getTableName().' '.        		
-        		"WHERE id ='".$this->getPrimaryKey()."'";
-       	$db->query($sql);
+		global $main, $db;
+		if ($main->checkToken()) {	
+			$sql = 'DELETE FROM '.$this->getTableName().' '.        		
+	        		"WHERE id ='".$this->getPrimaryKey()."'";
+	       	$db->query($sql);
+		}
 	}
 	/**
 	 * Gets the current table name
